Add option to invert the image (have y and i increase in the same direction).
I see images created in both ways (possibly due to a lack of consensus
around whether x, y images start at the top left or bottom left or who
knows why else), so let's make it a parameter so it can look like you
want.

Possibly will need to get renamed if we need to flip the x-axis as
well. We'll wait for that to happen.